Make codepoint.zig use u32 instead of u21 + more
Also adds lowerase checking and conversion for codepoints.

Previously we used u21 arrays and casted them to the u32 arrays that
grapheme expects. u21s in zig get padded to the next power of two, so
while the size was correct the value of the padding bytes is undefined.
For this reason the API now uses u32s for codepoint pointers. If you
have pointers to u21s and can ensure that the padding bits for each
codepoint are zero, then you can just cast them to a u32 slice to use
these functions.